A lot of css changes. Semantic footer back
Wow. I was really bad at using Zimit, and that's why I was frustrated.
Everything has to be under a row for it to get the column styling. Not
that hard, but for some reason that's why this project has taken so
long. Way to be Kevin.

Semantic footer is now back and sitting at the bottom without any extra
obscure corner of the internet css magic.